Terminado
El bug se corrige al utilizar Double como tipo de dato de salida del factorial. Long aumenta el limite de 17 a 20, pero a partir de 21 los números son muy largos. Double utiliza notación científica para los números problemáticos

+++ Repaso20/Repaso20/FactorialGet.cs
@@ -7,7 +7,7 @@ namespace Repaso20
         {
         }
 
-        public static int get(int numero)
+        public static double get(int numero)
         {
             
             if(numero == 1)
